+++ cmd/flagkey/consts.go
@@ -5,6 +5,7 @@ const (
 	Any         = "any"
 	Directory   = "dir"
 	IgnoreEmpty = "ignore"
+	Newline     = "newline"
 	Version     = "version"
 	Force       = "force"
 	Type        = "type"